pyproject.toml
[project]
name = "pruebapractica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.28.0",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
]
